packaging/sys-con: Only run postun steps when we are uninstalled.

+++ packaging/murphy.spec
@@ -561,10 +561,12 @@ ln -s %{_unitdir_user}/murphy-wait-for-launchpad-ready.path \
     %{_unitdir_user}/weston.target.wants/murphy-wait-for-launchpad-ready.path
 
 %postun system-controller
+if [ "$1" = "0" ]; then
 rm -f %{_unitdir_user}/weston.target.wants/murphy-wait-for-launchpad-ready.path
 if [ -f %{_unitdir_user}/ico-uxf-wait-launchpad-ready.path ]; then
     ln -sf %{_unitdir_user}/ico-uxf-wait-launchpad-ready.path \
         %{_unitdir_user}/weston.target.wants/ico-uxf-wait-launchpad-ready.path
+fi
 fi
 %endif
